WIP on splitting things up
